Added Confighelper for tax and added  an beginning of calculation of subtotal and tax to SalesViewModel

+++ SSCarlJohan.Desktop.UI.Library/Helpers/ConfigHelper.cs
@@ -5,11 +5,11 @@ namespace SSCarlJohan.Desktop.UI.Library.Helpers
 {
     public class ConfigHelper : IConfigHelper
     {
-        public double GetTaxRate()
+        public decimal GetTaxRate()
         {
             string rateText = ConfigurationManager.AppSettings["taxRate"];
 
-            bool isValidTaxRate = Double.TryParse(rateText, out double output);
+            bool isValidTaxRate = Decimal.TryParse(rateText, out decimal output);
 
             if (isValidTaxRate is false)
             {
